refactor: move filemanager from stateful to stateless
