+++ lib/workload/stateless/filemanager/database/Dockerfile
