+++ lib/workload/stateless/filemanager/database/queries/ingester/aws/insert_s3_created_objects.sql
