+++ lib/workload/stateless/filemanager/database/queries/ingester/aws/insert_s3_deleted_objects.sql
