+++ lib/workload/stateless/filemanager/database/queries/ingester/aws/update_deleted.sql
