+++ lib/workload/stateless/filemanager/database/queries/ingester/insert_objects.sql
