+++ lib/workload/stateless/filemanager/deploy/cdk.json
