+++ lib/workload/stateless/filemanager/docker-compose.yml
