+++ lib/workload/stateless/filemanager/filemanager-ingest-lambda/src/main.rs
