+++ lib/workload/stateless/filemanager/filemanager/src/clients/aws/config.rs
