+++ lib/workload/stateless/filemanager/filemanager/src/events/aws/collecter.rs
