+++ lib/workload/stateless/filemanager/filemanager/src/events/aws/collector_builder.rs
