+++ lib/workload/stateless/filemanager/filemanager/src/events/aws/mod.rs
