+++ lib/workload/stateless/filemanager/filemanager/src/handlers/aws.rs
