+++ lib/workload/stateless/filemanager/scripts/deploy.sh
