+++ lib/workload/stateless/filemanager/scripts/logs.sh
